requirements.txt
pandas>=2.0.0
pyodbc>=4.0.0
matplotlib>=3.5.0
pytest>=7.0.0
streamlit>=1.28.0
python-dotenv>=1.0.0
numpy>=1.26.0
altair>=5.0.0
plotly>=5.0.0
pytest-playwright>=0.4.0
